fix waves and random waves
You can now spawn both waves and random waves. Maybe I want to make it a little tighter, this could be achieved by spawning the random ones further up and picking a lower number as "last_in_wave" for the random spawns.

Also added wave 2

next I want to add start and end of a round so that the gameplay loop is complete