add wave 5 & 6 plus price tweaks
Wave 1-6 are now in the game

Also did some tweaking on costs in the shop so it's a bit more reasonable. Still need some sort of exponential curve to prices, linear doesn't really work. (since we get higher multiplier the longer the level is)

Made a demo as well. Feel like this was  a pretty good spot to do it. Kinda wen't back on the idea of making set levels and instead use random levels. I think I want to split it up into levels in the future, so that every level has a defined set of waves that are used for randomization